Move buffered parsers to their own package
