+++ pom.xml
@@ -38,6 +38,11 @@
       <artifactId>jfasta</artifactId>
       <version>2.1.1</version>
     </dependency>
+    <dependency>
+      <groupId>org.biojava</groupId>
+      <artifactId>biojava-core</artifactId>
+      <version>4.0.0</version>
+    </dependency>
   </dependencies>
 
   <build>
