+++ src/main/java/nl/tudelft/dnainator/parser/buffered/BufferedNodeParser.java
@@ -1,4 +1,4 @@
-package nl.tudelft.dnainator.parser;
+package nl.tudelft.dnainator.parser.buffered;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -8,6 +8,8 @@ import java.util.Map;
 
 import nl.tudelft.dnainator.core.Sequence;
 import nl.tudelft.dnainator.core.SequenceFactory;
+import nl.tudelft.dnainator.parser.InvalidHeaderFormatException;
+import nl.tudelft.dnainator.parser.NodeParser;
 
 /**
  * A variant of the {@link NodeParser}, which uses a {@link BufferedReader}.
